Cleaned up cities2 tileset and converted some cities
The new tileset adds a well, a double-wide bridge (vertical only)
and two new wall corners. I also wrote a conversion helper util
to make it easy to convert cities to the new tileset.

I cleaned up some of the cities, and started creating images for
them in docs/maps.  I still need to trim the images to not include
all the surrounding buffer tile space.