made screen size aware of windows scaling
Finally figured out what the heck was wrong. When I grab current_w and
current_h, that's the resolution the monitor is set at. But Windows can
also scale text and apps and such in the settings, and my game was
acting like it didn't know about that. This change checks if we are on a
Windows system and, if so, calls a dll function that makes it aware of
scaling. Then when I grab current_h and current_w it is modified
according to the scaling.