added overworld_map.tmx
This map is for a map viewing feature I will be putting in the
game. I'll have some kind of pause menu from which the player can
view a map of everywhere in the overworld he/she has been so far.
It has the places labeled, so that it will be easier for the
player to find places based on their names.