Add debug entry for map testing
This lets you start the game specifying a map and
coordinates where you want to start.  There is some
argument verification being done, but it's a debug
function, so it's not much.  It does not check if the
map you name actually exists.  It does not check if
the coordinates you specify are on a space that
is walkable. It's just a useful entry if you use
it right.

Start the game with
rotj.(py|sh) <map name> <x coord> <y coord>
for example:
./rotj.sh overworld 161 116