mark beaten path with melek at beginning of game
This ensures that when the game starts, if someone looks at the
pause map before exiting melek, they'll still see the area surrounding
melek. It's similar to how we mark the beaten path for an hq city
whenever an existing game loads, although the latter is only necessary
for backwards compatibility with older save files.