added ai_sprite_reference folder to quickly see ai_sprites to choose
This will help us as we decide in different cities and places which
ai sprites we want to use. We can just open up this folder in
data/images/ai_sprite_reference in a file explorer and preview
all the images with their names. The code does not use this folder
at all. Once we're done developing for the initial public release,
we should delete this folder as it won't be needed anymore to aid
development.