added sprites/stuff from bountiful to gid, fixed gid bug
The reason the game crashed every time we walked into gid was
because of some problem with the mountain tile. It was the first
city to use a mountain tile, and it couldn't handle it. I fixed it
by importing the tileset directly into gid.tmx instead of referencing
cities2.tsx. Not sure why it worked and the other way didn't, but
:shrug: