make sure portrait and menu box and stuff don't show at end of battle
When finishing a battle with a regular command (not all out or risk it),
the portrait of the current warlord taking a turn was still showing, as
well as a blank menu box or something like that. This was because we
needed to short circuit the execute_state being handled after a mini
move was handled and resulted in a win or a lose. Now the portrait and
erroneous menu or dialog boxes do not appear at the end of a battle.