Added compatibility with Wayland/Sway.
Bumped the version of all dependencies:
* python 3.8.5 -> 3.11.3
* pygame 2.0.1 -> 2.4.0
* pyscroll 2.19.2 -> 2.29.0
* PyTMX 3.21.7 -> 3.31.0
* PyInstaller 4.1 -> 5.13.0
* future 0.18.2 -> 0.18.3
* pywin32-ctypes 0.2.0 -> 0.2.0
* macholib 1.14 -> 1.16.2

The old version of pygame/SDL did not support Wayland correctly, but it was
enough to bump all the packages to fairly recent versions. Due to bug with
pygame/SDL, the attributes needed to make Wayland consider the pygame window
nonresizable (WindowMaxsize and WindowMinsize with equal values) are not
correctly set, leading to issues in Sway. The current workaround is to use
swaymsg to get the correct size of the window and reset the pygame display to
match, though this sort of finicky.

Also, for reasons not clear to me, the mixer has to be reinitialize to prevent
some errors relating to textboxes.