ROTJ doesn't crash when Corianton abandons the war
BUG: In a battle where Corianton abandons the war at the beginning, ROTJ would sometimes crash.
CAUSE: If an enemy tried to attack ally 5, but due to Corianton's abandonment there were only 4 allies, ROTJ crashed due to the indexing error
FIX: Remove Corianton from the list of allies before assigning index attributes to allies in a battle.

Since the index attribute was assigned before removing Corianton from the list of allies, in most cases some allies would always have index attributes that were incorrect, but only when the number of allies remaining was less than 5 could this result in an IndexError.